Reworked script logic
The script no longer provides an environment variable for calling
itself in order to access different menus. Menu files can now specify a
"men" field parallel to the "cmd" field (now optional) to indicate that
a different menu should be opened. The "gen" field should now be
specified inside an "items" list. Any "cmd" or "men" fields outside of
"items" are now only used in the case of invalid input or exiting the
menu.